+++ yolo/demos/video_detect_gpu.py
@@ -452,9 +452,9 @@ if __name__ == "__main__":
 
     #tf.train.Checkpoint.restore("/home/vishnu/Desktop/CAM2/TensorFlowModelGardeners/weights/weights/train_test_nojitter_helps_exit_early_1").expect_partial()
 
-    cap = FastVideo("test1.mp4",
+    cap = FastVideo("test4.mp4",
                     model="regular",
-                    model_version="v4",
+                    model_version="v3",
                     process_width=416,
                     process_height=416,
                     preprocess_with_gpu=True,
